pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "MCP多服务器客户端项目，支持JSON配置"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp-python",
    "fastmcp",
    "openai",
    "python-dotenv",
    "numpy>=2.2.4",
]
